Add  login and register page
