+++ app/containers/Auth/Login/index.js
@@ -1,12 +1,11 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import { AnimateField } from '../../../components/snappForm';
-import {loginSubmit,useFormInput} from './util';
+import { loginSubmit, useFormInput } from './util';
 import './style.scss';
-function Login(){
-    const loginUserName = useFormInput('');
-    const loginPass = useFormInput('');
-    const {SubmitComp,onKeyPress} = loginSubmit(loginUserName,loginPass);
+function Login() {
+  const loginUserName = useFormInput('john@jacob.com');
+  const loginPass = useFormInput('johnnyjacob');
+  const { SubmitComp, onKeyPress } = loginSubmit(loginUserName, loginPass);
   return (
     <div>
       <div className="loginForm">
@@ -28,12 +27,13 @@ function Login(){
           onChange={loginPass.onChange}
           value={loginPass.value}
           type="password"
-          label="222رمزعبور"
+          label="رمزعبور"
           onKeyPress={onKeyPress}
           required
         />
+
         <div className="topM40 wFull hP20 center">
-            <SubmitComp/>
+          <SubmitComp />
         </div>
       </div>
     </div>
